refactor: backend read
