+++ sessions/backend-read/demo-end/pages/api/jokes/[id].js
@@ -1,5 +1,5 @@
-import dbConnect from "../../../db/connect";
-import Joke from "../../../db/models/Joke";
+import dbConnect from "@/db/connect";
+import Joke from "@/db/models/Joke";
 
 export default async function handler(request, response) {
   await dbConnect();
@@ -9,9 +9,13 @@ export default async function handler(request, response) {
     const joke = await Joke.findById(id);
 
     if (!joke) {
-      return response.status(404).json({ status: "Not Found" });
+      response.status(404).json({ status: "Not Found" });
+      return;
     }
 
     response.status(200).json(joke);
+    return;
   }
+
+  response.status(405).json({ status: "Method not allowed." });
 }
