+++ sessions/backend-read/demo-start/pages/api/jokes/[id].js
@@ -1,4 +1,4 @@
-import { jokes } from "../../../lib/data.js";
+import { jokes } from "@/lib/data.js";
 
 export default function handler(request, response) {
   const { id } = request.query;
@@ -6,7 +6,8 @@ export default function handler(request, response) {
   const joke = jokes.find((joke) => joke.id === id);
 
   if (!joke) {
-    return response.status(404).json({ status: "Not Found" });
+    response.status(404).json({ status: "Not Found" });
+    return;
   }
 
   response.status(200).json(joke);
