+++ sessions/backend-read/demo-start/pages/api/jokes/index.js
@@ -1,4 +1,4 @@
-import { jokes } from "../../../lib/data.js";
+import { jokes } from "@/lib/data.js";
 
 export default function handler(request, response) {
   response.status(200).json(jokes);
