+++ sessions/backend-read/products/pages/index.js
@@ -1,10 +1,5 @@
-import ProductList from "../components/ProductList";
+import ProductList from "@/components/ProductList";
 import styled from "styled-components";
-
-const Heading = styled.h1`
-  text-align: center;
-  color: var(--color-nemo);
-`;
 
 export default function HomePage() {
   return (
@@ -19,3 +14,8 @@ export default function HomePage() {
     </>
   );
 }
+
+const Heading = styled.h1`
+  text-align: center;
+  color: var(--color-nemo);
+`;
